feat: :pushpin: validate  mongoose.Types.ObjectId all controller modules

+++ src/modules/combo/combo.controller.ts
@@ -21,6 +21,7 @@ import { uploadSingleImageInterceptor } from 'src/common/configs/upload';
 import { Request, Response } from 'express';
 import { Prisma } from '@prisma/client';
 import { PrismaDB } from '../prisma/prisma.extensions';
+import mongoose from 'mongoose';
 
 @UseInterceptors(uploadSingleImageInterceptor())
 @Controller('combo')
@@ -80,6 +81,8 @@ export class ComboController {
   }
   @Get('filter/:id')
   async findFilter(@Param('id') id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id))
+      return `not found mongoose Types ObjectId ${id}`;
     const combo = await this.comboService.findFilter(id);
     return combo;
   }
@@ -97,6 +100,8 @@ export class ComboController {
     @Res() res: Response,
   ) {
     try {
+      if (!mongoose.Types.ObjectId.isValid(id))
+        return `not found mongoose Types ObjectId ${id}`;
       const imgData = await this.uploadService.uploadSingleImageThirdParty(req);
       updateComboDto.picture = imgData.data.link;
       const combo = await this.comboService.update(id, updateComboDto);
@@ -111,6 +116,8 @@ export class ComboController {
 
   @Delete(':id')
   remove(@Param('id') id: string) {
+    if (!mongoose.Types.ObjectId.isValid(id))
+      return `not found mongoose Types ObjectId ${id}`;
     return this.comboService.remove(id);
   }
 }
